Variable numTruePV which stores getTrueNumInteractions has been changed from int to float cause getTrueNumInteractions() returns float.

+++ AnalysisCode/interface/YggdrasilEventVars.h
@@ -79,7 +79,7 @@ struct yggdrasilEventVars{
 
 
 
-  int     numTruePV_;
+  float   numTruePV_;
   int     numGenPV_;
   
   bool    GoodFirstPV_;
